--- frontend/src/assets/data.tsx
@@ -1,52 +1,0 @@
-const products = [
-    {
-        name: 'Nike Slim shirt',
-        id: 'nike-slim-shirt',
-        category: 'Shirts',
-        image: '/images/p1.png', // 679px × 829px
-        price: 120,
-        countInStock: 10,
-        brand: 'Nike',
-        rating: 4.5,
-        numReviews: 10,
-        description: 'high quality shirt',
-    },
-    {
-        name: 'Adidas Fit Shirt',
-        id: 'adidas-fit-shirt',
-        category: 'Shirts',
-        image: '/images/p2.png',
-        price: 250,
-        countInStock: 20,
-        brand: 'Adidas',
-        rating: 4.0,
-        numReviews: 10,
-        description: 'high quality product',
-    },
-    {
-        name: 'Nike Slim Pant',
-        id: 'nike-slim-pant',
-        category: 'Pants',
-        image: '/images/p3.png',
-        price: 25,
-        countInStock: 15,
-        brand: 'Nike',
-        rating: 4.5,
-        numReviews: 14,
-        description: 'high quality product',
-    },
-    {
-        name: 'Adidas Fit Pant',
-        id: 'adidas-fit-pant',
-        category: 'Pants',
-        image: '/images/p4.png',
-        price: 65,
-        countInStock: 5,
-        brand: 'Puma',
-        rating: 4.5,
-        numReviews: 10,
-        description: 'high quality product',
-    },
-]
-
-export default products;
