+++ frontend/src/components/Error.tsx
@@ -1,3 +1,5 @@
+// This component needs styling 
+// Material UI 
 import Alert from '@mui/material/Alert';
 import AlertTitle from '@mui/material/AlertTitle';
 import Stack from '@mui/material/Stack';
