+++ frontend/src/components/footer/Footer.tsx
@@ -1,8 +1,8 @@
-import React from 'react'
+// This component requires styling 
 
 const Footer = () => {
     return (
-        <div style={ { height: "30px", backgroundColor: "black", color: "white", textAlign: "center", position: "absolute", bottom: "0", width: "100%" } }>All rights reserved</div>
+        <div style={ { height: "30px", backgroundColor: "black", color: "white", textAlign: "center", width: "100%" } }>All rights reserved</div>
     )
 }
