+++ frontend/src/contexts/shopping-cart-context/shoppingCartContext.tsx
@@ -1,20 +1,14 @@
-import React, { createContext, useState } from "react";
-type shoppingCartChildren = {
-    children: React.ReactNode
-}
-type ShoppingCart = {
-    cartItems: product[]
-    setCartItems: React.Dispatch<React.SetStateAction<product[]>>
-}
-export type product = {
-    productId: string
-    quantity: number
-    productLimit: number
-}
+// react 
+import { createContext, useState } from "react";
+// types 
+import { product, ShoppingCart, shoppingCartChildren } from "./shoppingCartContextTypes";
+
+
 export const ShoppingCartContext = createContext({} as ShoppingCart);
 
 export const ShoppingCartProvider = ({ children }: shoppingCartChildren) => {
     const [cartItems, setCartItems] = useState<product[]>([])
+
     return (
 
         <ShoppingCartContext.Provider
