+++ frontend/src/fetchers/fetch.tsx
@@ -1,7 +1,10 @@
+// axios 
 import axios from "axios";
+// types 
 import { productsType } from "../components/display-products/displayProductsInterface";
 
 export const fetchProducts = async (URL: string): Promise<productsType[]> => {
+
     try {
         const { data } = await axios.get(URL);
         return await data
