+++ frontend/src/hooks/useProducts.tsx
@@ -1,10 +1,15 @@
+// useQuery 
 import { useQuery } from "react-query";
-import { fetchProducts } from "./fetch";
+// fetch 
+import { fetchProducts } from "../fetchers/fetch";
+
 const useProducts = (URL: string) => {
+
     const { isError, isLoading, data } = useQuery(["products"], () =>
         fetchProducts(URL)
     );
     return { isError, isLoading, data }
+
 }
 export default useProducts
