+++ frontend/src/utils/utils.tsx
@@ -1,35 +1,32 @@
+// Types 
 import { productsType } from "../components/display-products/displayProductsInterface"
-import { product } from "../contexts/shoppingCartContext"
+import { product } from "../contexts/shopping-cart-context/shoppingCartContextTypes"
 
-type props = {
+type addToShoppingCartTypes = {
     productName: string
     cartItems: product[]
     setCartItems: (value: product[]) => void
     product: productsType[]
 }
 
-export const addToShoppingCartLogic = ({ productName, cartItems, setCartItems, product }: props) => {
-    const find = cartItems.every(item => productName !== item.productId)
+export const addToShoppingCartLogic = ({ productName, cartItems, setCartItems, product }: addToShoppingCartTypes) => {
+    const productIsNotInShoppingCart = cartItems.every(item => productName !== item.productId)
 
-    if (find) {
+    if (productIsNotInShoppingCart) {
         return setCartItems([...cartItems, { productId: String(productName), quantity: 1, productLimit: 6 }])
-
     }
 
     return cartItems.forEach((item) => {
         if (item?.productId === productName && item?.productLimit <= 1) {
             return alert("product per purchase limit")
-
         }
 
         if (item?.productId === productName && product[0]?.countInStock <= item?.quantity) {
-
             return alert("error")
         }
 
         if (item?.productId === productName && product[0]?.countInStock >= item?.quantity) {
-
-            const filteredItem = cartItems.filter(item => {
+            const changeProductQuantity = cartItems.filter(item => {
                 if (item?.productId === productName) {
                     item.quantity += 1
                     item.productLimit -= 1
@@ -37,7 +34,7 @@ export const addToShoppingCartLogic = ({ productName, cartItems, setCartItems, p
                 }
                 return item
             })
-            return setCartItems([...filteredItem])
+            return setCartItems([...changeProductQuantity])
         }
     })
