+++ package.json
@@ -2,5 +2,9 @@
   "devDependencies": {
     "eslint": "^8.26.0",
     "eslint-plugin-react": "^7.31.10"
+  },
+  "dependencies": {
+    "@material-ui/core": "^4.12.4",
+    "@mui/styles": "^5.10.10"
   }
 }
